+++ src/snakes/GeneticProgramming/Subfunctions/ShortestPath.java
@@ -16,7 +16,7 @@ abstract class ShortestPath implements Subfunction {
      * @return distance to object
      */
     @Override
-    abstract public float value(Direction direction, Snake snake, Snake opponent, Coordinate mazeSize, Coordinate apple);
+    abstract public double value(Direction direction, Snake snake, Snake opponent, Coordinate mazeSize, Coordinate apple);
 
     protected float calculateShortestPath(Direction direction, Snake snake, Snake opponent, Coordinate mazeSize) {
         return 0;
